Experimenting with rotational script
From a bit i learned in the lynda course, I have written a script and
applied it as a Component to the Axis 1 Object of the ADJ Inno Spot
Model.  This script creates non-stop rotation for the correct Spot
light component around the Y axis.  It also features a variable exposed
the dev interface (Rate) which can change the behavior of the Script
realtime.